fix: correct syntax error
The code wasn't working because everything else was f64 expect one of the them. I replace the i32 with f64 then it started working again.

+++ src/main.rs
@@ -1,7 +1,7 @@
 fn main() {
     let price: f64 = 2.5;
     let quantity: f64 = 4.0;
-    let total: i32;
+    let total: f64;
     total = price * quantity;
 
     println!("Total: {total}");
